Untangle the packages
