+++ sdk/security/keyvault/azadmin/assets.json
@@ -2,5 +2,5 @@
   "AssetsRepo": "Azure/azure-sdk-assets",
   "AssetsRepoPrefixPath": "go",
   "TagPrefix": "go/security/keyvault/azadmin",
-  "Tag": "go/security/keyvault/azadmin_f576072351"
+  "Tag": "go/security/keyvault/azadmin_c42864e822"
 }
