+++ sdk/security/keyvault/azkeys/assets.json
@@ -2,5 +2,5 @@
   "AssetsRepo": "Azure/azure-sdk-assets",
   "AssetsRepoPrefixPath": "go",
   "TagPrefix": "go/security/keyvault/azkeys",
-  "Tag": "go/security/keyvault/azkeys_afd23684f8"
+  "Tag": "go/security/keyvault/azkeys_ff983e62e3"
 }
